mcp-web-app/requirements.txt
flask>=3.0.0
fastmcp>=2.8.0
python-dotenv>=1.0.0
uvicorn>=0.32.0
asgiref>=3.8.0
